Changes Made
Changed all of my spelling errors in the front-end.
All changes made in the front-end  has made my CRUD functionality functional.
Changed the Controller name in the employee-service file, for help, as there was a mis match between was was inputted and what appeared on swagger.
Added a notification to the view-employee .ts file using the snackbar function.
It works, however, a different approach will be further explored.
-Remember to change the path in the app-routing.ts file, for your edit function, in order to successfully retrieve the "ID" of a specific data record.